Safari-609.1.11.1/LayoutTests/fast/dom/HTMLTableRowElement/insertCell-default-argument.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/tabular-data.html#the-tr-element">
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Tests that HTMLTableRowElement.insertCell()'s default argument is -1.");

var tr = document.createElement("tr");
shouldBe("tr.__proto__", "HTMLTableRowElement.prototype");

tr.insertCell().innerHTML = "1"; // Should append.
tr.insertCell().innerHTML = "2"; // Should append.
tr.insertCell(-1).innerHTML = "3"; // Should append.
tr.insertCell(0).innerHTML = "0"; // Should prepend.

var cells = tr.cells;
shouldBe("cells.length", "4");
shouldBeEqualToString("cells[0].innerHTML", "0");
shouldBeEqualToString("cells[1].innerHTML", "1");
shouldBeEqualToString("cells[2].innerHTML", "2");
shouldBeEqualToString("cells[3].innerHTML", "3");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
